Description: 安全规范告警清理0428 IssueNo: https://gitee.com/openharmony/arkui_ace_engine/issues/I9KE5P Feature or Bugfix: Bugfix Binary Source:No Signed-off-by: chenjiafeng <chenjiafeng10@huawei.com>

+++ frameworks/bridge/declarative_frontend/jsview/js_image.cpp
@@ -652,7 +652,7 @@ void JSImage::SetEnhancedImageQuality(const JSCallbackInfo& info)
 void JSImage::CreateImageAnimation(std::vector<RefPtr<PixelMap>>& pixelMaps, int32_t duration, int32_t iterations)
 {
     std::vector<ImageProperties> imageList;
-    for (int i = 0; i < pixelMaps.size(); i++) {
+    for (int i = 0; i < static_cast<int32_t>(pixelMaps.size()); i++) {
         ImageProperties image;
         image.pixelMap = pixelMaps[i];
         imageList.push_back(image);
